backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.1

# Caching & Queue
redis==5.0.1
celery==5.3.4

# PDF Generation
reportlab==4.0.8

# Data Processing
pandas==2.1.3
numpy==1.26.2

# ML & Analytics
scikit-learn==1.3.2
statsmodels==0.14.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
